+++ __tests__/cli.spec.ts
@@ -9,10 +9,7 @@ const CLI_PATH = join(__dirname, '..')
 const projectName = 'test-app'
 const genPath = join(__dirname, projectName)
 
-const run = (
-  args: string[],
-  options: SyncOptions<string> = {}
-): ExecaSyncReturnValue<string> => {
+const run = (args: string[], options: SyncOptions<string> = {}): ExecaSyncReturnValue<string> => {
   return execaCommandSync(`node ${CLI_PATH} ${args.join(' ')}`, options)
 }
 
@@ -58,9 +55,7 @@ test('prompts for the framework on not supplying a value for --template', () => 
 
 test('prompts for the framework on supplying an invalid template', () => {
   const { stdout } = run([projectName, '--template', 'unknown'])
-  expect(stdout).toContain(
-    `"unknown" isn't a valid template. Please choose from below:`
-  )
+  expect(stdout).toContain(`"unknown" isn't a valid template. Please choose from below:`)
 })
 
 test('asks to overwrite non-empty target directory', () => {
@@ -77,7 +72,7 @@ test('asks to overwrite non-empty current directory', () => {
 
 test('successfully scaffolds a project based on vue starter template', () => {
   const { stdout } = run([projectName, '--template', 'vue'], {
-    cwd: __dirname
+    cwd: __dirname,
   })
   const generatedFiles = readdirSync(genPath).sort()
 
@@ -88,7 +83,7 @@ test('successfully scaffolds a project based on vue starter template', () => {
 
 test('works with the -t alias', () => {
   const { stdout } = run([projectName, '-t', 'vue'], {
-    cwd: __dirname
+    cwd: __dirname,
   })
   const generatedFiles = readdirSync(genPath).sort()