+++ template-react-ts/src/content/index.ts
@@ -1,0 +1,3 @@
+console.info('chrome-extension-boilerplate-vite-svelte content script')
+
+export {}