+++ template-react-ts/vite.config.ts
@@ -1,7 +1,21 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
+import { crx } from '@crxjs/vite-plugin'
+import manifest from './src/manifest'
 
 // https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [react()]
+export default defineConfig(({ mode }) => {
+  return {
+    build: {
+      emptyOutDir: true,
+      outDir: 'build',
+      rollupOptions: {
+        output: {
+          chunkFileNames: 'assets/chunk-[hash].js',
+        },
+      },
+    },
+
+    plugins: [crx({ manifest }), react()],
+  }
 })