+++ template-vue-js/src/popup/index.js
@@ -1,0 +1,5 @@
+import { createApp } from 'vue'
+import './style.css'
+import App from './Popup.vue'
+
+createApp(App).mount('#app')