+++ template-vue-ts/vite.config.ts
@@ -1,7 +1,24 @@
+import path from 'path'
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
+import { crx } from '@crxjs/vite-plugin'
+
+import manifest from './src/manifest'
 
 // https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [vue()]
+export default defineConfig(({ mode }) => {
+  const production = mode === 'production'
+
+  return {
+    build: {
+      emptyOutDir: true,
+      outDir: 'build',
+      rollupOptions: {
+        output: {
+          chunkFileNames: 'assets/chunk-[hash].js',
+        },
+      },
+    },
+    plugins: [crx({ manifest }), vue()],
+  }
 })